Changes made during our meeting up to 6:00 pm.  set up connection to the DB in the search controller.  Pushing to be able to merge form information
