+++ src/DatabaseSearch/SearchController.java
@@ -78,18 +78,7 @@ public class SearchController {
 
     // Connect to the DB
     protected Connection DBConnect() throws SQLException {
-        // I have no idea what this does
-        try {
-            Class.forName("com.mysql.jdbc.Driver").newInstance();
-        } catch (ClassNotFoundException cnfe) {
-            System.err.println("Error: " + cnfe.getMessage());
-        } catch (InstantiationException ie) {
-            System.err.println("Error: " + ie.getMessage());
-        } catch (IllegalAccessException iae) {
-            System.err.println("Error: " + iae.getMessage());
-        }
-        // ???
-        return DriverManager.getConnection(url, user, pass);
+        return TTB_database.connect();
     }
 
     // Function will query the DB
